refactor: 조회 기능 구현 위한 cqrs 적용
